style(daemon): better module separation
This crate exposes two sets of elements:
- the Zenoh-Flow daemon and how to create it,
- a set of Queries to interact and manage a Daemon from the outside

This commit sets up this logical separation.
